Photo nodes in order
On collision with each of the nodes, a different photo is shown to user. Press space or 'A' on xbox to continue. Node that needs to be collected next is orange, rest are blue. Can only be collected in the right order. Still needs a way to reset the level/prevent continuing if done in wrong order.